chore: Update npm dependencies and add Bootstrap and React Bootstrap
This commit updates the npm dependencies in the `package.json` file. It adds Bootstrap version 5.3.3 and React Bootstrap version 2.10.2 to the project. This will allow us to use the Bootstrap framework and its components in our React application.

The commit also updates the `index.js` file to remove unused imports and comments. It removes the import statements for `ReactDOM` from the `react-dom/client` module and the `index.css` file. It also removes the comment about measuring performance using `reportWebVitals`.

These changes will improve the performance and maintainability of our application by using the latest versions of the dependencies and removing unnecessary code.