chore: Update npm dependencies and add Bootstrap CSS
This commit updates the npm dependencies in the `package.json` file. It adds Bootstrap version 5.3.3 to the project and imports the Bootstrap CSS in the `index.js` file. This will allow us to use the Bootstrap framework and its styling in our React application.

These changes will improve the appearance and user experience of our application by leveraging the Bootstrap CSS.