chore: Update package.json scripts
The code changes in this commit update the scripts in the package.json file. The "start" script is modified to run the server and client concurrently using the "concurrently" package. The "server" script is added to run the backend server, and the "client" script is added to start the frontend client.

By updating the scripts, we improve the development workflow by easily running both the server and client simultaneously, making it more convenient to work on the application.