feat: Add data import and destroy scripts
The code changes include adding two new scripts, "data:import" and "data:destroy", to the package.json file. These scripts are used for importing and destroying data in the backend server.

By adding these scripts, we provide a convenient way to populate the database with initial data and also to remove all data when needed. This improves the development workflow and allows for easier testing and debugging.