refactor: Enable devTools in store.js
Enable the devTools option in the store.js file to allow for better debugging and monitoring of the application state. This change adds the "devTools: true" configuration to the store's middleware. It improves the development experience by providing access to the Redux DevTools extension for inspecting and manipulating the state.